+++ dubbo-metrics/dubbo-metrics-metadata/src/test/java/org/apache/dubbo/metrics/metadata/MetadataStatCompositeTest.java
@@ -17,7 +17,10 @@
 
 package org.apache.dubbo.metrics.metadata;
 
-import org.apache.dubbo.metrics.metadata.stat.MetadataStatComposite;
+import org.apache.dubbo.metrics.data.ApplicationStatComposite;
+import org.apache.dubbo.metrics.data.BaseStatComposite;
+import org.apache.dubbo.metrics.data.RtStatComposite;
+import org.apache.dubbo.metrics.data.ServiceStatComposite;
 import org.apache.dubbo.metrics.metadata.type.ApplicationType;
 import org.apache.dubbo.metrics.model.container.LongContainer;
 import org.junit.jupiter.api.Assertions;
@@ -27,15 +30,24 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
 
-import static org.apache.dubbo.metrics.metadata.MetadataMetricsConstants.OP_TYPE_SUBSCRIBE;
+import static org.apache.dubbo.metrics.metadata.MetadataMetricsConstants.*;
 
 public class MetadataStatCompositeTest {
 
     private final String applicationName = "app1";
 
+    private final BaseStatComposite statComposite = new BaseStatComposite() {
+        @Override
+        protected void init(ApplicationStatComposite applicationStatComposite, ServiceStatComposite
+            serviceStatComposite, RtStatComposite rtStatComposite) {
+            applicationStatComposite.init(MetadataMetricsConstants.appKeys);
+            serviceStatComposite.init(MetadataMetricsConstants.serviceKeys);
+            rtStatComposite.init(OP_TYPE_PUSH, OP_TYPE_SUBSCRIBE, OP_TYPE_STORE_PROVIDER_INTERFACE);
+        }
+    };
+
     @Test
     void testInit() {
-        MetadataStatComposite statComposite = new MetadataStatComposite();
         Assertions.assertEquals(statComposite.getApplicationStatComposite().getApplicationNumStats().size(), ApplicationType.values().length);
         //(rt)5 * (push,subscribe,service)3
         Assertions.assertEquals(5 * 3, statComposite.getRtStatComposite().getRtStats().size());
@@ -51,14 +63,12 @@ public class MetadataStatCompositeTest {
 
     @Test
     void testIncrement() {
-        MetadataStatComposite statComposite = new MetadataStatComposite();
-        statComposite.incrementApp(ApplicationType.P_TOTAL.getMetricsKey(), applicationName,1);
+        statComposite.incrementApp(ApplicationType.P_TOTAL.getMetricsKey(), applicationName, 1);
         Assertions.assertEquals(1L, statComposite.getApplicationStatComposite().getApplicationNumStats().get(ApplicationType.P_TOTAL.getMetricsKey()).get(applicationName).get());
     }
 
     @Test
     void testCalcRt() {
-        MetadataStatComposite statComposite = new MetadataStatComposite();
         statComposite.calcApplicationRt(applicationName, OP_TYPE_SUBSCRIBE.getType(), 10L);
         Assertions.assertTrue(statComposite.getRtStatComposite().getRtStats().stream().anyMatch(longContainer -> longContainer.specifyType(OP_TYPE_SUBSCRIBE.getType())));
         Optional<LongContainer<? extends Number>> subContainer = statComposite.getRtStatComposite().getRtStats().stream().filter(longContainer -> longContainer.specifyType(OP_TYPE_SUBSCRIBE.getType())).findFirst();
